github_project/requirements-dev.txt
astroid==2.0.4
backcall==0.1.0
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
decorator==4.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.2
djangorestframework==3.9.0
idna==2.7
ipdb==0.11
ipython==7.1.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.3.1
pep8==1.7.1
pickleshare==0.7.5
prompt-toolkit==2.0.7
Pygments==2.2.0
pylint==2.1.1
python-decouple==3.1
pytz==2018.7
requests==2.20.0
six==1.11.0
static3==0.7.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24
wcwidth==0.1.7
wrapt==1.10.11
